user registration added

+++ admin/includes/view_all_posts.php
@@ -64,7 +64,6 @@ if(isset($_POST['checkBoxArray'])){
             <th>Tags</th>
             <th>Comments</th>
             <th>Date</th>
-            <th>View</th>
         </tr>
     </thead>
     <tbody>
@@ -111,7 +110,7 @@ while ($row = mysqli_fetch_assoc($select_all_posts_admin)){
     echo "<td>$post_date</td>";
     echo "<td><a class='btn btn-info' href='../post.php?p_id={$post_id}'>View Post</a></td>";
     echo "<td><a class='btn btn-success' href='posts.php?source=edit_posts&p_id={$post_id}'>Edit</a></td>";
-    echo "<td><a class='btn btn-danger' href='posts.php?delete={$post_id}'>Delete</a></td>";
+    echo "<td><a class='btn btn-danger' onClick=\"javascript: return confirm('Are you sure you want to delete this post?') \"  href='posts.php?delete={$post_id}'>Delete</a></td>";
     echo "</tr>";
 ;}
